requirement.txt
openai==2.20.0
streamlit==1.54.0
google-generativeai